pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs", "grpcio-tools>=1.70.0"]
build-backend = "hatchling.build"

[tool.hatch.build.hooks.custom]
path = "build_proto.py"

[project]
name = "smolagents-crew"
dynamic = ["version"]
description = "A powerful framework for orchestrating multiple AI agents to work together efficiently"
authors = [{name = "Carlos André Antunes"}]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "smolagents",
    "grpcio>=1.70.0",
    "protobuf>=5.29.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov",
    "black",
    "isort",
    "flake8"
]
viz = [
    "networkx>=2.6.0",
    "matplotlib>=3.4.0"
]
all = [
    "networkx>=2.6.0",
    "matplotlib>=3.4.0",
    "pytest>=7.0.0",
    "pytest-cov",
    "black",
    "isort",
    "flake8"
]

[project.urls]
Homepage = "https://github.com/antunsz/smolagents-crew"

[project.scripts]
smolagents = "smolagents_crew.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["smolagents_crew"]
include = ["smolagents_crew/swarm/proto/*.proto"]

[tool.hatch.version]
source = "vcs"